# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Universo Agustino API
  version: 1.0.0
  description: API v1 - marketplace académico UNSA (comunidad @unsa.edu.pe). Compra en custodia sin comisión.
servers:
  - url: http://localhost:4000/api/v1
...
paths:
...
  /staff/shifts:
    get:
      tags: [staff]
      summary: Turnos del equipo para Técnico; solo propios para Trabajador
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: OK }, '201': { description: Creado }, '400': { description: VALIDATION }, '401': { description: Sin sesión }, '403': { description: Solo Técnico para cambios }, '404': { description: No encontrado }, '409': { description: Conflicto de horario, turno, sede o duplicado } }
    post:
      tags: [staff]
      summary: Crear turno dentro del horario, sede activa y sin solape entre sedes
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [userId, sedeId, weekday, startsMin, endsMin], properties: { userId: { type: string }, sedeId: { type: string }, weekday: { type: integer, minimum: 0, maximum: 6 }, startsMin: { type: integer, minimum: 0, maximum: 1440 }, endsMin: { type: integer, minimum: 0, maximum: 1440 } } }
      responses: { '200': { description: OK }, '201': { description: Creado }, '400': { description: VALIDATION }, '401': { description: Sin sesión }, '403': { description: Solo Técnico para cambios }, '404': { description: No encontrado }, '409': { description: Conflicto de horario, turno, sede o duplicado } }
  /staff/shifts/{id}:
    patch:
      tags: [staff]
      summary: Editar turno validando horario y solape (Técnico)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [userId, sedeId, weekday, startsMin, endsMin], properties: { userId: { type: string }, sedeId: { type: string }, weekday: { type: integer, minimum: 0, maximum: 6 }, startsMin: { type: integer, minimum: 0, maximum: 1440 }, endsMin: { type: integer, minimum: 0, maximum: 1440 } } }
      responses: { '200': { description: OK }, '201': { description: Creado }, '400': { description: VALIDATION }, '401': { description: Sin sesión }, '403': { description: Solo Técnico para cambios }, '404': { description: No encontrado }, '409': { description: Conflicto de horario, turno, sede o duplicado } }
    delete:
      tags: [staff]
      summary: Quitar turno (Técnico)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses: { '200': { description: OK }, '201': { description: Creado }, '400': { description: VALIDATION }, '401': { description: Sin sesión }, '403': { description: Solo Técnico para cambios }, '404': { description: No encontrado }, '409': { description: Conflicto de horario, turno, sede o duplicado } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...